+++ private/buf/cmd/buf/command/registry/module/modulelabel/modulelabellist/modulelabellist.go
@@ -172,7 +172,7 @@ func nextPageCommand(container appext.Container, flags *flags, nextPageToken str
 	if nextPageToken == "" {
 		return ""
 	}
-	command := fmt.Sprintf("buf registry label list %s", container.Arg(0))
+	command := fmt.Sprintf("buf registry module label list %s", container.Arg(0))
 	if flags.ArchiveStatus != bufcli.DefaultArchiveStatus {
 		command = fmt.Sprintf("%s --%s %s", command, archiveStatusName, flags.ArchiveStatus)
 	}